requirements.txt
opencensus-ext-azure
opencensus-ext-fastapi
fastapi
uvicorn
